cruft cleanup
